requirements-dev.txt
Faker==11.1.0
flake8==5.0.4
ipython==7.30.1
model-bakery==1.3.3

# Deps do pytest
attrs==20.3.0
py==1.10.0
pluggy==0.13.1
pytest-mock==3.6.1

#Coverage
coverage==6.2
codecov==2.1.12
pytest-cov==3.0.0
pytest-django==4.5.2


-r requirements.txt